+++ src/Exception/NotEmbeddableFileTypeException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\AssetsBundle\Exception;
+
+class NotEmbeddableFileTypeException extends AssetsException
+{
+}