+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *   Added support for modern + legacy builds for JS. These will only be linked if there are both a `@namespace/file.js` 
     as well as a `@namespace/_legacy.file.js` entry in the dependency map.
     These imports + their dependencies will be loaded using either `type="module"` (modern) or `nomodule` (legacy).
+*   Updated composer dependencies
 
 2.4.0
 =====